Return particle polarisations from `phase_vels`
Add the following to the items in the named tuple returned by
`phase_vels`:

- `:xp`
- `:xs1`
- `:xs2`

These are particle motions in the x1-x2-x3 coordinate system expressed
as unit vectors (and are just the eigenvectors of the Christoffel
matrix).

Update the docstring to reflect that and to make it clearer that a named
tuple is returned.